changed u32 to usize

+++ src/layer.rs
@@ -19,7 +19,7 @@ pub struct Layer {
 }
 
 impl Layer {
-    pub fn new(input_size: u32, output_size: u32, activation: ActivationType, seed: u64) -> Self {
+    pub fn new(input_size: usize, output_size: usize, activation: ActivationType, seed: u64) -> Self {
 
         // Xavier Initialization
         let scale = (2.0 / input_size as f32).sqrt();
